requirements.txt
openai
python-dotenv
flask
hdbcli
pandas
sqlalchemy
sqlalchemy-hana
rapidfuzz